recipe getting averageRating

+++ backend/src/main/java/com/ufscar/pooa/backend/dto/RecipeDTO.java
@@ -7,11 +7,11 @@ import jakarta.validation.constraints.NotNull;
 public record RecipeDTO (
     @NotBlank String name,
     @NotBlank String preparationMethods,
-    @NotNull Float rating,
+    @NotNull Double rating,
     @NotBlank List<String> ingredients,
     @NotBlank List<String> categories,
     List<String> comments) {
-      public RecipeDTO(String name, String preparationMethods, Float rating, List<String> ingredients, List<String> categories, List<String> comments) {
+      public RecipeDTO(String name, String preparationMethods, Double rating, List<String> ingredients, List<String> categories, List<String> comments) {
         this.name = name;
         this.preparationMethods = preparationMethods;
         this.rating = rating;
